Implement Justin's suggestion ,check eqlb of density in pymbar

+++ reproducibility_project/src/engines/mcccs/analysis.py
@@ -14,6 +14,11 @@ from scipy import stats
 def main():
     """Read run files from all the independent seeds and get the final density and RDFs."""
     data_path = "analysis_data"
+    if os.path.exists(data_path):
+        shutil.rmtree(data_path)
+    os.makedirs(data_path)
+    # delete the folder manually
+
     os.chdir(data_path)
 
     project = signac.get_project()
@@ -27,15 +32,26 @@ def main():
         )
     ):
         print(molecule, ensemble, temperature, pressure)
+        if not os.path.isdir(
+            "{}_{}_{}K_{}kPa".format(molecule, ensemble, temperature, pressure)
+        ):
+
+            os.makedirs(
+                "{}_{}_{}K_{}kPa".format(
+                    molecule, ensemble, temperature, pressure
+                )
+            )
         os.chdir(
             "{}_{}_{}K_{}kPa".format(molecule, ensemble, temperature, pressure)
         )
+
         base_dir = os.getcwd()
         if ensemble == "NPT" and molecule == "methaneUA":
 
             density_list = []
             for job in group:
                 os.chdir(job.ws)
+                #        print(job)
                 prod_run_files = sorted(glob("run*prod*"))
                 if len(prod_run_files) < 4:
                     print(
@@ -49,7 +65,8 @@ def main():
                         )
                     )
                 density_list.append(avg_one_seed_density(prod_run_files))
-            print(len(traj_list))
+
+            #    print(density_list)
             filtered_density_list = list(filter(None, density_list))
             output_string = "The average density is {} g/ml with SEM {} from {} samples".format(
                 np.mean(filtered_density_list),
@@ -63,8 +80,8 @@ def main():
             text_file.write(output_string)
             text_file.close()
 
-            comb_traj = md.load_hdf5("comb_traj.h5")
-            print(comb_traj)
+            # comb_traj = md.load_hdf5("comb_traj.h5")
+            # print(comb_traj)
         elif ensemble == "GEMC-NVT" and molecule == "methaneUA":
             for job in group:
                 print(job)
@@ -77,12 +94,15 @@ def avg_one_seed_density(prod_run_files):
     if len(prod_run_files) == 0:
         return None
     os.system(
-        "grep 'specific density                        ' run.prod* | awk '{print $5'} > temp_density.txt"
+        "grep 'specific density                        ' run.prod* | awk '{print $6'} > temp_density.txt"
     )
-    density_list_one_seed = np.genfromtxt("temp_density.txt")
-    # now delete the temp file
-    os.remove("temp_density.txt")
-    return np.mean(density_list_one_seed)
+    if os.path.exists("temp_density.txt"):
+        density_list_one_seed = np.genfromtxt("temp_density.txt")
+        # now delete the temp file
+        os.remove("temp_density.txt")
+        return np.mean(density_list_one_seed)
+    else:
+        return None
 
 
 if __name__ == "__main__":
